// src/verb-types.ts
export type GodanVerb = {
  type: "godan";
  stem: string;
  ending: "う" | "く" | "ぐ" | "す" | "つ" | "ぬ" | "ぶ" | "む" | "る";
};

export type IchidanVerb = {
  type: "ichidan";
  stem: string;
  ending: "る";
};

export type IrregularVerb = {
  type: "irregular";
  dictionary: "する" | "来る" | "くる";
};